add assetURI tests, cleanup files
