+++ ranger/tmpl/layout/authenticated_base.tmpl
@@ -14,7 +14,7 @@
     {{ template "additionalHeadScripts" }}
 
     {{ if not isDevelopment }}
-    <link rel="stylesheet" href="{{asset "assets/style.css"}}">
+    <link rel="stylesheet" href="{{ asset "assets/style.css" }}">
     {{ end }}
   </head>
